extract LinkViewHolder from Link model
